✅ Componentização da CheckBox em DelSystem
